+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 Change:
 
-- Added __sharedasset__ as a placeholder for the h_user_id field, allowing an asset to be set as shared.
+- Added __sharedasset__ as a placeholder for the h_used_by field, allowing an asset to be set as shared.
 
 Fixed:
 
-- modification to re-introduce [HBAssetType] as a placeholder - now introduced as __hbassettype__ - for the Asset Type. 
+- modification to re-introduce [HBAssetType] as a placeholder - now introduced as __hbassettype__ - for the Asset Type. Note that this is case-sensitive. 
 
 ## 2.2.0 (August 5th, 2021)